requirements.txt
aiogram
pywin32
Pillow
python-dotenv
aiohttp
aiofiles